Decompose util to multiple files
