+++ src/lists.js
@@ -1,4 +1,6 @@
-import {Events, genId, defer} from "./util";
+import Events from "./util/events";
+import genId from "./util/gen-id";
+import defer from "./util/deferred";
 import {router} from "./router";
 
 const db = firebase.database();
